Move SidePanel component into Panels directory
