+++ client/galaxy/scripts/components/Panels/SidePanel.vue
